upscaling/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pillow==11.1.0
python-multipart==0.0.20
opencv-python==4.11.0.86
basicsr==1.4.2
facexlib==0.3.0
gfpgan==1.3.8
numpy<2
realesrgan==0.3.0
--index-url https://download.pytorch.org/whl/cu118
torch==2.0.0
torchvision==0.15.0
# git+https://github.com/xinntao/Real-ESRGAN.git
# git+https://github.com/TencentARC/GFPGAN.git